requirements.txt
asgiref==3.4.1
coverage==6.0.1
Django==3.2.5
pytz==2021.3
sqlparse==0.4.2
libgravatar==1.0.0
django-widget-tweaks==1.4.8
Faker==8.10.3
scikit-surprise==1.1.1
scipy==1.8.0
joblib==1.1.0
numpy==1.22.3
pandas==1.4.1
gunicorn
django-heroku

humanize==3.12.0
python-dateutil==2.8.2
six==1.16.0
text-unidecode==1.3
